opt some

+++ index.sh
@@ -9,6 +9,7 @@ _temp_file() { # 加载文件 url arg...
     ctx_temp=$(mktemp) && _down_file $ctx_temp $1 && shift && source $ctx_temp "$@"
 }
 _down_tar() { # 下载文件 dir url
+    [ -f $1 ] && return
     echo "download $ctx_dev/$2"
     curl --create-dirs -o $1 -fSL $ctx_dev/$2 && tar -xvf $1
 }
@@ -22,10 +23,11 @@ prepare_system() {
     esac
 }
 prepare_package() {
-    prepare_system
-    local back=$PWD; cd ~/
+    prepare_system; local back=$PWD; cd ~/
     _down_tar go.tar.gz publish/go.tar.gz
     _down_tar vim.tar.gz publish/vim.tar.gz
+    _down_tar ish.tar.gz publish/ish.tar.gz
+    _down_tar local.bin.tar.gz publish/local.bin.tar.gz
     cd $back
 }
 prepare_script() {
@@ -71,7 +73,7 @@ main() {
             ;;
         binary) # 应用安装
             export PATH=${PWD}/bin:$PATH ctx_log=${ctx_log:=/dev/stdout}
-            shift && prepare_ice && bin/ice.sh serve serve start dev dev "$@"
+            shift && prepare_ice && bin/ice.sh serve serve start "$@"
             ;;
         dev) # 开发环境
             prepare_package; prepare_script plug.sh conf.sh miss.sh; ish_sys_dev_config
